requirements.txt
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
django-filter==23.3